fw/kernel/util/stop: make stop mode no-op on nRF5
The whole architecture of the stop mode is flawed. This pseudo-generic
API is in reality a way to disable certain sleep modes on STM32, where
you can't enter some of them without breaking some peripherals state.
The proper way to fix this would be to have an API that platform
specific drivers use when peripherals are being used, explicitly
mentioning which platform/soc specific LP mode is not allowed.

The current API is being used in e.g. generic layers of I2C, where it
does not make any sense in platforms like nRF5. Just one more example of
PebbleOS sloppiness in its low level _architecture_.

For now, just make this a no-op on nRF5. This will hopefully be
refactored before everything explodes.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>