bluetooth-fw/nimble: adjust in progress flag setting
bt_driver_cb_gatt_client_discovery_complete could end up calling
bt_driver_gatt_stop_discovery on timeouts, so if in progress flag was
still set, the stop call would stall on a semaphore forever. The whole
logic of this needs to be reviewed, for now this _should_ fix the
observed problems.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>