kernel/stop: turn off dbgserial UART during stop on power-hungry platforms
On nRF52, it is very expensive to keep the UART turned on -- it forces the
PCLK to stay active.  So, if we're going to enter stop mode, switch it off
on the way in.  We'll get woken by an EXTI anyway -- and even if we lose the
first byte as a result, since we're speaking PULSE, the LCP-UP command will
be retransmitted (PULSE is a reliable protocol!) and we'll get woken into
stop-inhibited mode in time to catch the rest of it.

Tested with ./waf console and with ./waf image_resources while inspecting
current.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>